split into GC independent/dependent code
