+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/notification/GcNotificationRequest.java
@@ -24,7 +24,7 @@
  * questions.
  */
 
-package com.oracle.svm.core.genscavenge.service;
+package com.oracle.svm.core.notification;
 
 import com.oracle.svm.core.heap.GCCause;
 import org.graalvm.nativeimage.Platform;
