+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/notification/Target_com_sun_management_internal_GcInfoBuilder.java
@@ -24,7 +24,7 @@
  * questions.
  */
 
-package com.oracle.svm.core.genscavenge.service;
+package com.oracle.svm.core.notification;
 
 import com.oracle.svm.core.annotate.Alias;
 import com.oracle.svm.core.annotate.Substitute;
